+++ services/resources/mgmt/2015-11-01/resources/models.go
@@ -464,6 +464,47 @@ type DeploymentPropertiesExtended struct {
 	ParametersLink *ParametersLink `json:"parametersLink,omitempty"`
 	// Mode - Gets or sets the deployment mode. Possible values include: 'Incremental', 'Complete'
 	Mode DeploymentMode `json:"mode,omitempty"`
+	// Error - READ-ONLY; The deployment error.
+	Error *ErrorResponse `json:"error,omitempty"`
+}
+
+// MarshalJSON is the custom marshaler for DeploymentPropertiesExtended.
+func (dpe DeploymentPropertiesExtended) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]interface{})
+	if dpe.ProvisioningState != nil {
+		objectMap["provisioningState"] = dpe.ProvisioningState
+	}
+	if dpe.CorrelationID != nil {
+		objectMap["correlationId"] = dpe.CorrelationID
+	}
+	if dpe.Timestamp != nil {
+		objectMap["timestamp"] = dpe.Timestamp
+	}
+	if dpe.Outputs != nil {
+		objectMap["outputs"] = dpe.Outputs
+	}
+	if dpe.Providers != nil {
+		objectMap["providers"] = dpe.Providers
+	}
+	if dpe.Dependencies != nil {
+		objectMap["dependencies"] = dpe.Dependencies
+	}
+	if dpe.Template != nil {
+		objectMap["template"] = dpe.Template
+	}
+	if dpe.TemplateLink != nil {
+		objectMap["templateLink"] = dpe.TemplateLink
+	}
+	if dpe.Parameters != nil {
+		objectMap["parameters"] = dpe.Parameters
+	}
+	if dpe.ParametersLink != nil {
+		objectMap["parametersLink"] = dpe.ParametersLink
+	}
+	if dpe.Mode != "" {
+		objectMap["mode"] = dpe.Mode
+	}
+	return json.Marshal(objectMap)
 }
 
 // DeploymentsCreateOrUpdateFuture an abstraction for monitoring and retrieving the results of a