+++ services/resources/mgmt/2015-11-01/subscriptions/subscriptionsapi/interfaces.go
@@ -20,6 +20,7 @@ var _ BaseClientAPI = (*subscriptions.BaseClient)(nil)
 
 // ClientAPI contains the set of methods on the Client type.
 type ClientAPI interface {
+	CheckZonePeers(ctx context.Context, subscriptionID string, parameters subscriptions.CheckZonePeersRequest) (result subscriptions.CheckZonePeersResult, err error)
 	Get(ctx context.Context, subscriptionID string) (result subscriptions.Subscription, err error)
 	List(ctx context.Context) (result subscriptions.ListResultPage, err error)
 	ListComplete(ctx context.Context) (result subscriptions.ListResultIterator, err error)