+++ services/resources/mgmt/2019-05-01/resources/models.go
@@ -708,6 +708,8 @@ type DeploymentPropertiesExtended struct {
 	DebugSetting *DebugSetting `json:"debugSetting,omitempty"`
 	// OnErrorDeployment - The deployment on error behavior.
 	OnErrorDeployment *OnErrorDeploymentExtended `json:"onErrorDeployment,omitempty"`
+	// Error - READ-ONLY; The deployment error.
+	Error *ErrorResponse `json:"error,omitempty"`
 }
 
 // MarshalJSON is the custom marshaler for DeploymentPropertiesExtended.
@@ -2189,7 +2191,8 @@ type ProviderResourceType struct {
 	// Aliases - The aliases that are supported by this resource type.
 	Aliases *[]AliasType `json:"aliases,omitempty"`
 	// APIVersions - The API version.
-	APIVersions *[]string `json:"apiVersions,omitempty"`
+	APIVersions  *[]string      `json:"apiVersions,omitempty"`
+	ZoneMappings *[]ZoneMapping `json:"zoneMappings,omitempty"`
 	// Capabilities - The additional capabilities offered by this resource type.
 	Capabilities *string `json:"capabilities,omitempty"`
 	// Properties - The properties.
@@ -2210,6 +2213,9 @@ func (prt ProviderResourceType) MarshalJSON() ([]byte, error) {
 	}
 	if prt.APIVersions != nil {
 		objectMap["apiVersions"] = prt.APIVersions
+	}
+	if prt.ZoneMappings != nil {
+		objectMap["zoneMappings"] = prt.ZoneMappings
 	}
 	if prt.Capabilities != nil {
 		objectMap["capabilities"] = prt.Capabilities
@@ -2642,4 +2648,11 @@ func (future *ValidateMoveResourcesFuture) result(client Client) (ar autorest.Re
 	}
 	ar.Response = future.Response()
 	return
+}
+
+// ZoneMapping ...
+type ZoneMapping struct {
+	// Location - The location of the zone mapping.
+	Location *string   `json:"location,omitempty"`
+	Zones    *[]string `json:"zones,omitempty"`
 }