+++ services/resources/mgmt/2015-11-01/resources/CHANGELOG.md
@@ -1,2 +1,13 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### New Funcs
+
+1. DeploymentPropertiesExtended.MarshalJSON() ([]byte, error)
+
+### Struct Changes
+
+#### New Struct Fields
+
+1. DeploymentPropertiesExtended.Error