+++ services/resources/mgmt/2016-07-01/resources/CHANGELOG.md
@@ -1,2 +1,18 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### New Funcs
+
+1. DeploymentPropertiesExtended.MarshalJSON() ([]byte, error)
+
+### Struct Changes
+
+#### New Structs
+
+1. ZoneMapping
+
+#### New Struct Fields
+
+1. DeploymentPropertiesExtended.Error
+1. ProviderResourceType.ZoneMappings