+++ services/resources/mgmt/2018-05-01/resources/CHANGELOG.md
@@ -1,2 +1,14 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### Struct Changes
+
+#### New Structs
+
+1. ZoneMapping
+
+#### New Struct Fields
+
+1. DeploymentPropertiesExtended.Error
+1. ProviderResourceType.ZoneMappings